Split prepareFull into its own file
